refactor(comments): arrange components
